app-harmonia/backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Albano Maywitz",email = "allbano@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.131.0,<0.132.0)",
    "edwh-uuid7 (>=0.2.3,<0.3.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pyjwt (>=2.8.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.2,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
puthonpath = 'app'
addopts = '-p no:warnings'
testpaths = ['tests']

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app/main.py'
run2 = 'uvicorn app.main:app --reload --port 55030'
pre_test = 'task lint'
test = 'pytest -s -x  --cov=app -vv'
post_test = 'coverage html'
